+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/firFactoryUtil.kt
@@ -6,6 +6,7 @@ import dev.zacsweers.metro.compiler.fir.FirTypeKey
 import dev.zacsweers.metro.compiler.fir.Keys
 import dev.zacsweers.metro.compiler.fir.MetroFirValueParameter
 import dev.zacsweers.metro.compiler.fir.buildFullSubstitutionMap
+import dev.zacsweers.metro.compiler.fir.buildHiddenFromObjCAnnotation
 import dev.zacsweers.metro.compiler.fir.buildSimpleValueParameter
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.compatContext
@@ -13,6 +14,7 @@ import dev.zacsweers.metro.compiler.fir.copyParameters
 import dev.zacsweers.metro.compiler.fir.generateMemberFunction
 import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.fir.metroFirBuiltIns
+import dev.zacsweers.metro.compiler.fir.replaceAnnotationsSafe
 import dev.zacsweers.metro.compiler.fir.wrapInProviderIfNecessary
 import dev.zacsweers.metro.compiler.ir.IrTypeKey
 import dev.zacsweers.metro.compiler.symbols.Symbols
@@ -44,7 +46,6 @@ import org.jetbrains.kotlin.fir.types.ConeClassLikeType
 import org.jetbrains.kotlin.fir.types.ConeKotlinType
 import org.jetbrains.kotlin.fir.types.constructType
 import org.jetbrains.kotlin.name.CallableId
-import org.jetbrains.kotlin.name.Name
 
 internal fun FirDeclarationGenerationExtension.buildFactoryConstructor(
   context: MemberGenerationContext,
@@ -107,22 +108,6 @@ internal fun FirDeclarationGenerationExtension.buildFactoryConstructor(
       }
       .also { it.containingClassForStaticMemberAttr = owner.toLookupTag() }
   }
-
-internal fun FirDeclarationGenerationExtension.buildFactoryCreateFunction(
-  context: MemberGenerationContext,
-  returnType: ConeKotlinType,
-  instanceReceiver: ConeClassLikeType?,
-  extensionReceiver: ConeClassLikeType?,
-  valueParameters: List<MetroFirValueParameter>,
-): FirNamedFunctionSymbol {
-  return buildFactoryCreateFunction(
-    context,
-    { returnType },
-    instanceReceiver,
-    extensionReceiver,
-    valueParameters,
-  )
-}
 
 @OptIn(SymbolInternals::class)
 internal fun FirDeclarationGenerationExtension.buildFactoryCreateFunction(
@@ -208,78 +193,9 @@ internal fun FirDeclarationGenerationExtension.buildFactoryCreateFunction(
               .toFirResolvedTypeRef()
         }
       }
-      .symbol as FirNamedFunctionSymbol
-  }
-
-@OptIn(SymbolInternals::class)
-internal fun FirDeclarationGenerationExtension.buildNewInstanceFunction(
-  context: MemberGenerationContext,
-  name: Name,
-  returnType: ConeKotlinType,
-  instanceReceiver: ConeClassLikeType?,
-  extensionReceiver: ConeClassLikeType?,
-  valueParameters: List<MetroFirValueParameter>,
-): FirNamedFunctionSymbol =
-  with(session.compatContext) {
-    return generateMemberFunction(
-        context.owner,
-        returnType.toFirResolvedTypeRef(),
-        CallableId(context.owner.classId, name),
-        origin = Keys.FactoryNewInstanceFunction.origin,
-      ) {
-        val thisFunctionSymbol = symbol
-
-        val containingClassSymbol = context.owner.getContainingClassSymbol()!!
-        val ownerToCopyTypeParametersFrom: FirClassSymbol<*> =
-          if (context.owner.isCompanion) {
-            // companion -> class factory -> original class
-            containingClassSymbol.getContainingClassSymbol()!!
-          } else {
-            // object factory -> original class
-            containingClassSymbol
-          }
-            as FirClassSymbol<*>
-
-        val classTypeArgsToReplace = mutableMapOf<FirTypeParameterSymbol, ConeKotlinType>()
-        for (typeParameter in ownerToCopyTypeParametersFrom.typeParameterSymbols) {
-          typeParameters +=
-            buildTypeParameterCopy(typeParameter.fir) {
-                origin = Keys.Default.origin
-                this.symbol = FirTypeParameterSymbol()
-                containingDeclarationSymbol = thisFunctionSymbol
-              }
-              .also { classTypeArgsToReplace[typeParameter] = it.symbol.constructType() }
-        }
-
-        instanceReceiver?.let {
-          this.valueParameters +=
-            buildSimpleValueParameter(
-              name = Symbols.Names.instance,
-              type = it.toFirResolvedTypeRef(),
-              containingFunctionSymbol = thisFunctionSymbol,
-              origin = Keys.InstanceParameter.origin,
-            )
-        }
-        extensionReceiver?.let {
-          this.valueParameters +=
-            buildSimpleValueParameter(
-              name = Symbols.Names.receiver,
-              type = it.toFirResolvedTypeRef(),
-              containingFunctionSymbol = thisFunctionSymbol,
-              origin = Keys.ReceiverParameter.origin,
-            )
-        }
-
-        copyParameters(
-          functionBuilder = this,
-          sourceParameters = valueParameters,
-          // Will be copied in IR
-          copyParameterDefaults = false,
-        ) { original ->
-          val type = original.contextKey.originalType(session)
-          val substitutor = substitutorByMap(classTypeArgsToReplace, session)
-          val copiedType = substitutor.substituteOrNull(type) ?: type
-          this.returnTypeRef = copiedType.toFirResolvedTypeRef()
+      .also { func ->
+        buildHiddenFromObjCAnnotation(session)?.let {
+          func.replaceAnnotationsSafe(func.annotations + it)
         }
       }
       .symbol as FirNamedFunctionSymbol