+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/ir.kt
@@ -1440,6 +1440,19 @@ internal fun buildAnnotation(
   }
 }
 
+/**
+ * Adds `@HiddenFromObjC` to [function] if the annotation is available (K/N only).
+ *
+ * We do this because there's a bunch of places where K/N linking breaks on generated symbols.
+ *
+ * https://github.com/ZacSweers/metro/issues/2137
+ */
+context(context: IrMetroContext)
+internal fun addHiddenFromObjCAnnotation(function: IrFunction) {
+  val ctor = context.metroSymbols.hiddenFromObjCAnnotationConstructor ?: return
+  function.annotations += buildAnnotation(function.symbol, ctor)
+}
+
 internal val IrClass.metroGraphOrFail: IrClass
   get() = metroGraphOrNull ?: reportCompilerBug("No generated MetroGraph found: $classId")
 
@@ -1606,6 +1619,16 @@ internal val IrFunction.isNamedCopy: Boolean
 
 private val COMPONENT_FUNCTION_REGEX = Regex(StandardNames.DATA_CLASS_COMPONENT_PREFIX + "[0-9]*")
 
+// private fun isComponentNMethod(method: CallableMemberDescriptor): Boolean {
+//    if ((method as? FunctionDescriptor)?.isOperator != true) return false
+//    val parent = method.containingDeclaration
+//    if (parent is ClassDescriptor && parent.isData &&
+// DataClassResolver.isComponentLike(method.name)) {
+//        // componentN method of data class.
+//        return true
+//    }
+//    return false
+// }
 internal val IrFunction.isComponentOperator: Boolean
   get() {
     return this is IrSimpleFunction &&