+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/symbols/Symbols.kt
@@ -440,6 +440,10 @@ internal class Symbols(
     pluginContext.referenceClass(ClassIds.ComptimeOnly)?.constructors?.first()!!
   }
 
+  val hiddenFromObjCAnnotationConstructor: IrConstructorSymbol? by lazy {
+    pluginContext.referenceClass(ClassIds.HiddenFromObjC)?.constructors?.first()
+  }
+
   val throwsAnnotationConstructor: IrConstructorSymbol? by lazy {
     // For some reason this isn't visible until 2.3.0?
     pluginContext.referenceClass(ClassIds.Throws)?.constructors?.first()